+++ python/GafferSceneUI/__init__.py
@@ -215,6 +215,7 @@ from . import CameraQueryUI
 from . import ReflectionConstraintUI
 from . import CurvesInterpolationUI
 from . import MeshLightUI
+from . import AttributeProcessorUI
 
 # then all the PathPreviewWidgets. note that the order
 # of import controls the order of display.